Fix fixture
